% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_calc_annual.R
\name{clim_calc_annual}
\alias{clim_calc_annual}
\title{clim_calc_annual}
\usage{
clim_calc_annual(
  site,
  parameter,
  start_year,
  end_year,
  select_year,
  water_year_start
)
}
\value{
A tibble of annual climate data
}
\description{
Returns annual values of selected climate data
}
