% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_calc_daily.R
\name{clim_calc_daily}
\alias{clim_calc_daily}
\title{clim_calc_daily}
\usage{
clim_calc_daily(
  site,
  parameter,
  start_year,
  end_year,
  select_year,
  water_year_start
)
}
\value{
A tibble of daily climate data
}
\description{
Returns daily values of selected climate data in a tibble suitable to calculate monthly and annual summary data
}
