% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_compile_daily.R
\name{clim_compile_daily}
\alias{clim_compile_daily}
\title{clim_compile_daily}
\usage{
clim_compile_daily(
  site,
  parameter,
  start_year,
  end_year,
  select_year,
  rain_cutoff = 0
)
}
\value{
A tibble of daily climate data
}
\description{
Returns daily values of selected climate data
}
