% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_plot_dayofyear.R
\name{clim_plot_dayofyear}
\alias{clim_plot_dayofyear}
\title{clim_plot_dayofyear}
\usage{
clim_plot_dayofyear(
  site = "Fort Simpson",
  parameter = "mean_temp",
  start_year = 1981,
  end_year = 2024,
  select_year = 2022
)
}
\value{
A ggplot2 object of climate data
}
\description{
Returns a boxplot of climate data for each day of the year
}
