% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_plot_monthly.R
\name{clim_plot_monthly}
\alias{clim_plot_monthly}
\title{clim_plot_monthly}
\usage{
clim_plot_monthly(
  site,
  parameter,
  select_year,
  water_year_start = 1,
  water_year_end = NA,
  start_year = 1950,
  end_year = 2024,
  max_missing_days = 3,
  y_min = NA,
  y_max = NA,
  select_year_point_size = 2,
  historic_point_size = 1,
  legend_position = c(0.1, 0.95),
  save = FALSE,
  plot_width = 16,
  plot_height = 10,
  dpi = 900,
  file_name = "Default climate plot",
  extension = "png"
)
}
\value{
A ggplot2 object of monthly climate data
}
\description{
Returns a boxplot of monthly climate data
}
