% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_plot_to_date.R
\name{clim_plot_to_date}
\alias{clim_plot_to_date}
\title{clim_plot_to_date}
\usage{
clim_plot_to_date(
  site = c("Mackenzie"),
  parameter = "SWE",
  select_year = lubridate::year(Sys.Date()),
  water_year_start = 10,
  start_year = 1950,
  end_year = 2024,
  end_date = Sys.Date(),
  max_missing_days = 15,
  y_min = NA,
  y_max = NA,
  plot_break = FALSE,
  select_year_point_size = 2,
  historic_point_size = 1,
  legend_position = c(0.1, 0.95),
  save = FALSE,
  plot_width = 18,
  plot_height = 11,
  dpi = 900,
  file_name = "Default climate plot",
  extension = "png"
)
}
\value{
A ggplot2 object of climate data
}
\description{
Returns a boxplot of climate data up to a certain date
}
