% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_calc_daily.R
\name{hydro_calc_daily}
\alias{hydro_calc_daily}
\title{hydro_calc_daily}
\usage{
hydro_calc_daily(
  station_number,
  parameter,
  water_year_start = 1,
  start_date = "1900-01-01",
  end_date = Sys.Date(),
  timezone = "America/Edmonton",
  realtime_dl = T
)
}
\value{
A tibble of the daily values
}
\description{
Calculates daily values of hydrometric data, along with summary stats for each day of year
}
