% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_calc_dayofyear.R
\name{hydro_calc_dayofyear}
\alias{hydro_calc_dayofyear}
\title{hydro_calc_dayofyear}
\usage{
hydro_calc_dayofyear(
  station_number,
  parameter,
  select_years = lubridate::year(Sys.Date()),
  water_year_start = 1,
  after_bennett = FALSE,
  historic_min = NA,
  historic_max = NA
)
}
\value{
A tibble of the daily values
}
\description{
Calculates daily values of hydrometric data, along with summary stats for each day of year
}
