% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_compile_daily.R
\name{hydro_compile_daily}
\alias{hydro_compile_daily}
\title{hydro_compile_daily}
\usage{
hydro_compile_daily(
  station_number,
  parameter,
  start_date = "1900-01-01",
  end_date = Sys.Date(),
  timezone = "America/Edmonton",
  realtime_dl = T
)
}
\value{
A tibble of the daily values
}
\description{
Returns daily values of selected hydrometric data
}
