% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_plot_daily.R
\name{hydro_plot_daily}
\alias{hydro_plot_daily}
\title{hydro_plot_daily}
\usage{
hydro_plot_daily(
  station_number,
  parameter = "level",
  select_years = lubridate::year(Sys.Date()),
  after_bennett = FALSE,
  historic_min = NA,
  historic_max = lubridate::year(Sys.Date()) - 1,
  water_year_start = 1,
  historic = TRUE,
  log_scale = FALSE,
  start_month = 1,
  start_day = 1,
  end_month = 12,
  end_day = 31,
  line_colours = c("blue4", "orange4", "green4", "red4", "purple4", "yellow4"),
  legend_position = "top",
  line_size = 0.5,
  point_size = 0,
  legend_text_size = 8,
  y_min = NA,
  y_max = NA,
  save = FALSE,
  plot_width = 18,
  plot_height = 11,
  dpi = 900,
  file_name = "Default hydrometric plot",
  extension = "png"
)
}
\value{
A ggplot object of the selected station
}
\description{
Plots daily values of hydrometric data relative to historic values for each day of year
}
