% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFun.R
\name{testFun}
\alias{testFun}
\title{This is a test function}
\usage{
testFun(x)
}
\arguments{
\item{x}{[numeric] value or vector}
}
\value{
Returns \code{x} * 100
}
\description{
This function multiplies a numeric value (or vector) by 100. I am adding additional
information here to make sure this works. Does this go \strong{bold}?
Is this \emph{italics}?
Does this link to this function in this pakage \code{\link{xyPlot}}.
Email is: \email{shaun.p.harrigan@gmail.com} (note the doubled @).
Web is: \url{http://rstudio.com}. Here is a list:
\enumerate{
  \item First item in list
  \item Second item in list
}
End of description.
}
\examples{
# Simple
testFun(10)
# More complex
testFun(2345.33)

}
\seealso{
\code{\link{xyPlot}}
}
\keyword{test}

