% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyPlot.R
\name{xyPlot}
\alias{xyPlot}
\title{Simple xy plot}
\usage{
xyPlot(x, y, title = "Simple plot")
}
\arguments{
\item{x:}{numeric vector (e.g. Years)}

\item{y:}{numeric vector (e.g. Flow)}

\item{title:}{[string] Title for plot}
}
\value{
A plot object
}
\description{
Simple line xy plot, with red line.
}
\examples{
xyPlot()
}
\seealso{
\code{\link{testFun}}
}
\keyword{Plot}

