% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyPlotBlue.R
\name{xyPlotBlue}
\alias{xyPlotBlue}
\title{Simple xy plot}
\usage{
xyPlotBlue(x, y, title = "Simple plot - Blue")
}
\arguments{
\item{x:}{numeric vector (e.g. Years)}

\item{y:}{numeric vector (e.g. Flow)}

\item{title:}{[string] Title for plot}
}
\value{
A plot object
}
\description{
Simple line xy plot, with red line.
}
\examples{
xyPlotBlue()
}
\seealso{
\code{\link{testFun}}
}
\keyword{Plot}

